refactor : SingleContentUiState 확장함수 처리 #PAR-353

+++ feature/running/src/main/java/online/partyrun/partyrunapplication/feature/running/single/SingleContentScreen.kt
@@ -276,7 +276,7 @@ fun RunningBackNavigationHandler(
 
 fun handleRunningExit(
     activity: Activity?,
-    distanceInMeter: Int,
+    distanceInMeter: Double,
     singleContentViewModel: SingleContentViewModel,
     minimumFinishDistance: Int
 ) {
